fix: Main.tsx 로직을 hooks로 분리
